+++ app/Domain/Variation/Services/VariationService.php
@@ -7,7 +7,6 @@ use App\Domain\Variation\Models\Variation;
 use App\Domain\Variation\Models\VariationType;
 use App\Domain\Variation\Models\VariationTypeValue;
 use App\Http\Media\Request\StoreMediaRequest;
-use App\Http\Variation\Requests\StoreVariationRequest;
 use App\Http\Variation\Resources\VariationResource;
 use App\Http\Variation\Resources\VariationTypeResource;
 use App\Http\Variation\Resources\VariationTypeValueResource;
@@ -17,6 +16,7 @@ use App\Support\Services\Media\ImageService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class VariationService
 {
@@ -32,15 +32,13 @@ class VariationService
 
     public function getVariationDetails(Variation $variation): JsonResponse|VariationResource
     {
-        if (auth()->user()->store()->first()->id === $variation->store_id) {
-            return new VariationResource(
-                $variation->load(['children' => function ($query) {
-                    $query->with('variationType', 'variationTypeValue');
-                }, 'VariationColor', 'VariationImages', 'variationType', 'variationTypeValue'])
-            );
-        } else {
-            return response()->json([], 401);
-        }
+        Gate::authorize('view', $variation);
+
+        return new VariationResource(
+            $variation->load(['children' => function ($query) {
+                $query->with('variationType', 'variationTypeValue');
+            }, 'VariationColor', 'VariationImages', 'variationType', 'variationTypeValue'])
+        );
     }
 
     public function getVariationTypes(): AnonymousResourceCollection
@@ -50,6 +48,11 @@ class VariationService
         );
     }
 
+    public function getVariationColorValues(): AnonymousResourceCollection
+    {
+        $color = VariationType::query()->with('variationTypeValues')->whereRaw("JSON_EXTRACT(type, '$.en') = 'color'")->first();
+        return VariationTypeValueResource::collection($color->variationTypeValues);
+    }
 
     public function getVariationTypeValues(): AnonymousResourceCollection
     {
@@ -59,31 +62,34 @@ class VariationService
     }
 
     /**
-     * @param StoreVariationRequest $request
+     * @param array $data
      * @param ImageService $imageService
      * @return mixed
      */
 
-    public function store(StoreVariationRequest $request, ImageService $imageService): void
+    public function store(array $data, ImageService $imageService): void
     {
-        $data = $request->validated();
+        $product = Product::query()
+            ->find($data['product_id']);
+
+        $isAdmin = Auth::guard('admin')->check();
+        $isOwner = $isAdmin ? false : auth()->user()->isOwner($product->store_id);
+
+        if (!$isAdmin && !$isOwner) abort(403);
 
         $variationType = VariationType::find($data['variation_type_id']);
         $variationTypeValue = VariationTypeValue::find($data['variation_type_value_id']);
 
         $data['is_stockable'] = $variationType->is_stockable;
 
-
-        $product = Product::query()
-            ->find($data['product_id']);
-
         $data['title'] = $this->formVariationTitle($variationTypeValue->value, $product->title);
+
 
         $variation = $product->variations()
             ->create($data);
 
-        if ($request->hasFile('images') && $variationType->is_mediable && $variation) {
-            $imageService->imageUpload($variation, 'images', MediaCollectionEnums::VARIATION, $variation->id, ['primary' => false]);
+        if (isset($data['images']) && $variationType->is_mediable && $variation) {
+            $imageService->imageUpload($variation, 'images', MediaCollectionEnums::VARIATION, $variation->id);
         }
     }
 
@@ -92,8 +98,67 @@ class VariationService
         return $typeValue . ' ' . $productTitle;
     }
 
+    public function createColorVariation(array $data, ImageService $imageService): VariationResource
+    {
+        $product = Product::query()
+            ->find($data['product_id']);
+
+        if (!auth()->user()->isOwner($product->store_id) && !Auth::guard('admin')->check()) abort(403);
+
+        $colorType = VariationType::query()->whereRaw("JSON_EXTRACT(type, '$.en') = 'color'")->first();
+
+        $variationTypeValue = VariationTypeValue::find($data['variation_type_value_id']);
+
+        if ($variationTypeValue->variation_type_id !== $colorType->id) abort(400);
+
+        $data['is_stockable'] = $colorType->is_stockable;
+
+        $data['title'] = $this->formVariationTitle($variationTypeValue->value, $product->title);
+
+        $data['variation_type_id'] = $colorType->id;
+
+
+        $variation = $product->variations()
+            ->create($data);
+
+        if (isset($data['images']) && $variation) {
+            $imageService->imageUpload($variation, 'images', MediaCollectionEnums::VARIATION, $variation->id);
+        }
+
+        return new VariationResource($variation->load(['variationSmallImage', 'variationTypeValue', 'variationType']));
+    }
+
+    public function createSizeVariation(array $data, ImageService $imageService): void
+    {
+        $product = Product::query()
+            ->find($data['product_id']);
+
+        if (!auth()->user()->isOwner($product->store_id) && !Auth::guard('admin')->check()) abort(403);
+
+        $sizeType = VariationType::query()->whereRaw("JSON_EXTRACT(type, '$.en') = 'size'")->first();
+
+        $variationTypeValue = VariationTypeValue::find($data['variation_type_value_id']);
+
+        if ($variationTypeValue->variation_type_id !== $sizeType->id) abort(400);
+
+        $data['is_stockable'] = $sizeType->is_stockable;
+
+        $data['variation_type_id'] = $sizeType->id;
+
+        $data['title'] = $this->formVariationTitle($variationTypeValue->value, $product->title);
+
+        $variation = $product->variations()
+            ->create($data);
+
+        if (isset($data['images']) && $variation) {
+            $imageService->imageUpload($variation, 'images', MediaCollectionEnums::VARIATION, $variation->id);
+        }
+    }
+
     public function update(array $data, Variation $variation, ImageService $imageService)
     {
+        Gate::authorize('update', $variation);
+
         $variationType = VariationType::query()->find($data['variation_type_id']);
 
         $data['is_stockable'] = $variationType->is_stockable;
@@ -108,6 +173,7 @@ class VariationService
     public function destroy(int $id)
     {
         $variation = Variation::query()->find($id);
+        Gate::authorize('delete', $variation);
         $variation?->delete();
     }
 
@@ -125,7 +191,7 @@ class VariationService
 
     public function addImagesToVariation(Variation $variation, StoreMediaRequest $request, ImageService $imageService)
     {
-        if (Auth::guard('admin')->check() || \auth()->user()->store()->first()->id === $variation->store_id) {
+        if (Auth::guard('admin')->check() || \auth()->user()->isOwner($variation->store_id)) {
             if ($request->hasFile('images')) {
                 $imageService->imageUpload($variation, 'images', MediaCollectionEnums::VARIATION, $variation->id);
             }
